update User module and dependencies
