+++ .eslintrc
@@ -30,7 +30,7 @@
     "settings":{
         "import/resolver": {
             "node": {
-                "extensions": [".js", ".ts"]
+                "extensions": [".js", ".jsx", ".ts", ".tsx"]
             }
         }
     }
