+++ package.json
@@ -30,6 +30,7 @@
     },
     "dependencies": {
         "@prisma/client": "^5.8.1",
-        "express": "^4.18.2"
+        "express": "^4.18.2",
+        "zod": "^3.22.4"
     }
 }
