+++ prisma/schema.prisma
@@ -11,10 +11,10 @@ datasource db {
 }
 
 model User {
-  id    String  @id @default(uuid())
-  email String  @unique
-  name  String?
-
+  id         String   @id @default(uuid())
+  email      String   @unique
+  name       String?
+  password   String
   created_at DateTime @default(now())
   updated_at DateTime @updatedAt
 }
