+++ src/modules/Auth/authRouter.ts
@@ -1,5 +1,5 @@
 import { Router } from 'express';
-import { authController } from './controller/auth.controller';
+import { authController } from './controller/authController';
 
 const router = Router();
 const baseUrl = '/auth';
