+++ src/modules/User/controller/userController.ts
@@ -1,0 +1,27 @@
+import { Request, Response } from 'express';
+import { z } from 'zod';
+
+import { userService } from '../service/userService';
+
+class UserController {
+    public async create(req: Request, res: Response) {
+        const { name, email, password } = req.body;
+        try {
+            const zUserSchema = z.object({
+                name: z.string().optional(),
+                email: z.string().email({ message: 'E-mail é obrigatório.' }),
+                password: z.string().min(8, { message: 'Senha é obrigatório.' }),
+            });
+            zUserSchema.parse({ name, email, password });
+        } catch (err) {
+            return res.status(400).json({
+                message: 'Dados inválido',
+                error: err,
+            });
+        }
+
+        await userService.create(name, email, password);
+    }
+}
+
+export const userContoller = new UserController();
