+++ src/modules/User/service/userService.ts
@@ -1,0 +1,17 @@
+import { prismaConnect } from '../../../prismaConn';
+
+class UserService {
+    public async create(name: string, email: string, password: string) {
+        const findUser = await prismaConnect.user.findUnique({
+            where: {
+                email,
+            },
+        });
+
+        if (findUser) {
+            throw new Error(`Usuário ${name} já existente: ${email}`);
+        }
+    }
+}
+
+export const userService = new UserService();
