+++ src/modules/User/userRouter.ts
@@ -1,0 +1,9 @@
+import { Router } from 'express';
+import { userContoller } from './controller/userController';
+
+const router = Router();
+const baseUrl = '/user';
+
+router.post(`${baseUrl}`, userContoller.create);
+
+export const userRouter = router;
